加入匯出測試的部份。
git-svn-id: http://svn.gwis.com.tw:88/repos/M_Charge_Project/trunk@5295 4b8841f4-42b5-4263-839e-868dbe4382dd

+++ src/main/java/com/gateweb/charge/repository/CashMasterRepository.java
@@ -34,7 +34,7 @@ public interface CashMasterRepository extends JpaRepository<CashMasterEntity, In
 
 	public List<CashMasterEntity> findByOutYm(String yearMonth);
 
-	public CashMasterEntity findByCashMasterId(Long cashMasterId);
+	public CashMasterEntity findByCashMasterId(Integer cashMasterId);
 
 	/*
